cleanup notebooks
